Modify installFeatures.sh script
Adding a while loop around each call to the opendaylight
karaf client. If the client is not available it will try again up
to 10 times. The new version of opendaylight seems to have delays after
installing sdnc components which can cause errors for the appc components.
These script changes fix allow the install script to wait for opendaylight to
become available.

Change-Id: I31d7e365fd528d015f1d4d07a7f67f55c058d1d0
Signed-off-by: Patrick Brady <pb071s@att.com>