Update patch set 2
Patch Set 2:

> Marcus, I fixed the error I made by mistake on the topology word,
 > thanks for the heads up. I tried making the commit msg two lines
 > instead of one but it threw an error that a msg could not exceed
 > more than 72 chars so i left it as is.

Thanks Hector. The gerrit GUI is not super helpful in editing, easier and recommended to do it locally with a 'git commit --amend' and then update the commit message, and push again. Thanks for trying, will approve this for now as it's minor issue - will submit when build succeeds.

Patch-set: 2