Add appc version properties
When we increment the version of the deployment project, builds
will currently break since the version will be one ahead of the
released version of appc.

Adding properties to define the version of appc to download.
Also adding a script to determine when snapshot or release version
should be used.

Change-Id: I50f6c8c64c5e80893a91810b6bff82609e5a36a1
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: CIMAN-120