Add appc version properties
(cherry-pick from master)

When we increment the version of the deployment project, builds
will currently break since the version will be one ahead of the
released version of appc.

Adding properties to define the version of appc to download.
Also adding a script to determine when snapshot or release version
should be used.

Change-Id: I8e98dbc8ca79e0ec6a02fb011c17053c728c351d
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: CIMAN-120