Temporary fix for jenkins
Temporary fix for the issue where maven running on jenkins
is not able to find the platform logic installer zip file
in the maven local repository. By setting this version to
snapshot, maven will be able to find the version in nexus
that was build by another job, since it can't find the one
being build by the docker build job.

Change-Id: I9e2125134f9ef0984eb3406848c98d6687212771
Signed-off-by: Patrick Brady <patrick.brady@att.com>
Issue-ID: APPC-1816